Fix a race condition within test
This race appears because some prints have been removed:
1. The TB listen & accept
2. The net-tester connects
3. The TB parses the net-tester outputs and sees that it
   is ready, then it checks its own socket to see if it
   is opened, and if not, try to connect to the net-tester
   as it assumes that in that case the net-tester is client.
4. The TB's on_accept callback (from 1) is finally executed
   while a connect involving the same socket is in flight (3)

=> Already Open error

The fix is a change to the step (3) to connect only when
the TB is acting as a server (i.e. do not rely on the
TB's socket state anymore as it is subject to the race
between the processing of the net-tester output and the TB's
on_accept callback).

Signed-off-by: David Keller <david.keller@enyx.fr>